refactoring project structure to match a more decoupled logic away from blacksheep framework. most of these layers do not have to be tied to a framework's folder. any other framework can simply consume these layers as well
