+++ xcov19/services/geolocation.py
@@ -1,0 +1,68 @@
+from __future__ import annotations
+
+import abc
+from typing import TypeVar, Protocol, Callable, List
+
+from xcov19.dto import LocationQueryJSON, Address, FacilitiesResult
+from xcov19.utils.mixins import InterfaceProtocolCheckMixin
+
+T = TypeVar("T", bound=LocationQueryJSON)
+
+
+class LocationQueryServiceInterface[T: LocationQueryJSON](Protocol):
+    """Location aware service for listing faciltiies.
+    1. Searches and fetches existing processed results by query_id for a cust_id
+    2. Resolves coordinates from a given geolocation.
+    3. Fetches all facilities from a given set of records for a
+    given radius from geolocation.
+
+    Radius is default for now.
+    # TODO: Filter to be added
+    """
+
+    @classmethod
+    @abc.abstractmethod
+    async def resolve_coordinates(
+        cls, reverse_geo_lookup_svc: Callable[[T], dict], query: T
+    ) -> Address:
+        raise NotImplementedError
+
+    @classmethod
+    @abc.abstractmethod
+    async def fetch_facilities(
+        cls,
+        reverse_geo_lookup_svc: Callable[[T], dict],
+        patient_query_lookup_svc: Callable[[Address, T], List[FacilitiesResult]],
+        query: T,
+    ) -> List[FacilitiesResult] | None:
+        raise NotImplementedError
+
+
+class GeolocationQueryService(
+    LocationQueryServiceInterface, InterfaceProtocolCheckMixin
+):
+    @classmethod
+    async def resolve_coordinates(
+        cls,
+        reverse_geo_lookup_svc: Callable[[LocationQueryJSON], dict],
+        query: LocationQueryJSON,
+    ) -> Address:
+        """Resolves to address by geo reverse lookup."""
+        return Address(**reverse_geo_lookup_svc(query))
+
+    @classmethod
+    async def fetch_facilities(
+        cls,
+        reverse_geo_lookup_svc: Callable[[LocationQueryJSON], dict],
+        patient_query_lookup_svc: Callable[
+            [Address, LocationQueryJSON], List[FacilitiesResult]
+        ],
+        query: LocationQueryJSON,
+    ) -> List[FacilitiesResult] | None:
+        """Fetches facilities for a query location for a query id for a customer."""
+        patient_address = await cls.resolve_coordinates(reverse_geo_lookup_svc, query)
+        return patient_query_lookup_svc(patient_address, query) or None
+
+
+# TODO: Implement reverse_geo_lookup_svc
+# TODO: Implement following method in patient_query_lookup_svc:
